replace auth_provider (type-erased) util fns w/ AuthProvider
