+++ crates/core/src/eth_engine/mod.rs
@@ -1,5 +1,7 @@
+mod auth_provider;
 mod auth_transport;
 mod util;
 mod valid_payload;
 
+pub use auth_provider::AuthProvider;
 pub use util::*;
